+++ src/modules/basket-modal/BasketItem/BasketItem.component.tsx
